backend/requirements.txt
# TexPro AI - Textile Manufacturing Optimization System
# CMDT (Compagnie Malienne pour le Développement des Textiles)
# Requirements file for MVP backend

# Core Django Framework
Django==5.2.5
djangorestframework==3.16.1

# Authentication & Security
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0

# API Filtering & Search
django-filter==24.3

# Database
# SQLite is included with Python (for MVP)
# For production PostgreSQL:
# psycopg2-binary==2.9.10

# Image Processing (for quality control photos)
Pillow==11.0.0

# Data Processing & Analytics
pandas==2.2.3
numpy==2.1.3

# Excel/CSV Export for Reports
openpyxl==3.1.5
xlsxwriter==3.2.0
reportlab==4.2.5

# Environment Management
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.2

# Timezone handling
pytz==2024.2

# Date utilities
python-dateutil==2.9.0

# HTTP requests (for external APIs)
requests==2.32.3

# Task scheduling (for maintenance predictions)
celery==5.4.0
redis==5.1.1

# File handling
python-magic==0.4.27

# Validation
jsonschema==4.23.0

# Testing Framework
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
faker==30.3.0

# Code Quality & Linting
flake8==7.1.1
black==24.8.0
isort==5.13.2
pre-commit==4.0.1

# Development Tools
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Monitoring & Logging
sentry-sdk==2.17.0

# Performance
django-cache-machine==1.2.0

# Email (for notifications)
django-anymail==12.0

# Backup utilities
django-dbbackup==4.2.0
